refactor: auth client generation
